backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pandas>=2.0.0
boto3>=1.29.0
python-dotenv>=1.0.0
pydantic[email]>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
xgboost>=1.7.0
motor>=3.3.2
pymongo>=4.6.1
bcrypt>=4.1.2
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.18
email-validator>=2.1.0
certifi
dnspython


